Remove activities from web/app
