--- packages/services/api/src/modules/activity/module.graphql.mappers.ts
@@ -1,4 +1,0 @@
-import type { ActivityObject } from '../../shared/entities';
-
-export type ActivityConnectionMapper = readonly ActivityObject[];
-export type ActivityMapper = ActivityObject;
