--- packages/services/api/src/modules/activity/resolvers/MemberAddedActivity.ts
@@ -1,7 +1,0 @@
-import type { MemberAddedActivityResolvers } from './../../../__generated__/types.next';
-
-export const MemberAddedActivity: MemberAddedActivityResolvers = {
-  __isTypeOf(activity) {
-    return activity.type === 'MEMBER_ADDED';
-  },
-};
