--- packages/services/api/src/modules/activity/resolvers/OrganizationCreatedActivity.ts
@@ -1,7 +1,0 @@
-import type { OrganizationCreatedActivityResolvers } from './../../../__generated__/types.next';
-
-export const OrganizationCreatedActivity: OrganizationCreatedActivityResolvers = {
-  __isTypeOf(activity) {
-    return activity.type === 'ORGANIZATION_CREATED';
-  },
-};
