--- packages/services/api/src/modules/activity/resolvers/ProjectCreatedActivity.ts
@@ -1,7 +1,0 @@
-import type { ProjectCreatedActivityResolvers } from './../../../__generated__/types.next';
-
-export const ProjectCreatedActivity: ProjectCreatedActivityResolvers = {
-  __isTypeOf(activity) {
-    return activity.type === 'PROJECT_CREATED';
-  },
-};
