--- packages/services/api/src/modules/activity/resolvers/TargetCreatedActivity.ts
@@ -1,7 +1,0 @@
-import type { TargetCreatedActivityResolvers } from './../../../__generated__/types.next';
-
-export const TargetCreatedActivity: TargetCreatedActivityResolvers = {
-  __isTypeOf(activity) {
-    return activity.type === 'TARGET_CREATED';
-  },
-};
