+++ packages/services/api/src/modules/project/providers/project-manager.ts
@@ -2,10 +2,10 @@ import { Injectable, Scope } from 'graphql-modules';
 import { paramCase } from 'param-case';
 import type { Project, ProjectType } from '../../../shared/entities';
 import { share, uuid } from '../../../shared/helpers';
-import { ActivityManager } from '../../activity/providers/activity-manager';
 import { AuthManager } from '../../auth/providers/auth-manager';
 import { OrganizationAccessScope } from '../../auth/providers/organization-access';
 import { ProjectAccessScope } from '../../auth/providers/project-access';
+import { ActivityManager } from '../../shared/providers/activity-manager';
 import { Logger } from '../../shared/providers/logger';
 import { OrganizationSelector, ProjectSelector, Storage } from '../../shared/providers/storage';
 import { TokenStorage } from '../../token/providers/token-storage';
